Add 'off' to SQL*Plus keywords
Even in SQL*Plus 'set' commands, the 'on' keyword is colorized according to the ANSI SQL join predicate (keyword.other.sql.oracle scope). I have added 'off' to the SQL*Plus keywords because, as far as I know, this is the only place the keyword is used, though it would be reason to include 'off' within the keyword.other.sql.oracle for consistency with 'set' and 'on'.